abstracted styles
