+++ src/client/pages/Landing/WhatWeDoCard.js
@@ -7,79 +7,7 @@ import useWindowSize from 'hooks/useWindowSize'
 
 import { connect } from 'react-redux'
 import { push } from 'react-router-redux'
-
-const getStyles = (theme, width, align) => {
-  let titleSize = theme.typography.h4
-  let textSize = theme.typography.body1
-  let imageSize = 100
-
-  console.log(theme.typography)
-  if (width > theme.breakpoints.values.sm) {
-    titleSize = theme.typography.h2
-    textSize = theme.typography.h6
-    imageSize = 150
-  }
-
-  if (width > theme.breakpoints.values.md) {
-    imageSize = 300
-  }
-
-  const curr = {
-    Image: {
-      width: imageSize,
-      height: imageSize,
-    },
-    ImageGrid: {
-      display: 'flex',
-      alignItems: 'center',
-      justifyContent: 'center',
-      backgroundColor: '#E9EBED',
-    },
-    Card: {
-      display: 'flex',
-      flexDirection: 'row',
-      flexFlow: 'row',
-      alignItems: 'stretch',
-      width: '86vw',
-      height: '37vh',
-      margin: '0 auto 2em',
-      padding: 10,
-      backgroundColor: 'white',
-    },
-    Title: {
-      ...titleSize,
-      color: '#D63333',
-    },
-    Text: {
-      ...textSize,
-      color: '#333333',
-      margin: '10px 0px',
-    },
-    Item: {
-      justifyContent: 'center',
-      display: 'flex',
-      flexDirection: 'column',
-      alignItems: 'flex-end',
-      textAlign: 'right',
-      padding: 25,
-    },
-    Tag: {
-      ...textSize,
-      color: '#D63333',
-      textTransform: 'uppercase',
-    },
-    Container: {
-      justifyContent: 'space-between',
-    },
-  }
-
-  if (align !== 'left') {
-    delete curr.Item.alignItems
-    delete curr.Item.textAlign
-  }
-
-  return curr
-}
+import getStyles from './WhatWeDoCard.styles'
 
 function WhatWeDoCard({ content, align, image, navigateTo }) {
   const theme = useTheme()
@@ -113,6 +41,7 @@ WhatWeDoCard.propTypes = {
   content: PropTypes.shape({}),
   align: PropTypes.string,
   image: PropTypes.string,
+  navigateTo: PropTypes.func.isRequired,
 }
 
 const mapDispatchToProps = (dispatch) => ({
